tests/gtrsim.dat

#data
&gtrsim
#errors
#document
| <html>
|   <head>
|   <body>
|     ">rsim"

#data
&gt
#errors
#document
| <html>
|   <head>
|   <body>
|     ">"

#data
<h2>&g</h2>
#errors
#document
| <html>
|   <head>
|   <body>
|     <h2>
|       "&g"
